third_party/blink/web_tests/platform/win11-arm64/virtual/dbsc-enabled/external/wpt/device-bound-session-credentials/include-site.https-expected.txt

This is a testharness.js-based test.
[FAIL] An established session refreshes across origins if the site is included
  assert_equals: expected 200 but got 401
Harness: the test ran to completion.
